refactor(producer): rename football status processor + extract shared base
EventCompetitionStatusDocumentProcessor (no sport prefix) was actually
football-only despite the misleading name; user confused themselves
twice in code review during PR #428 thinking the unprefixed name meant
something it didn't. Renamed to FootballEventCompetitionStatusDocumentProcessor
and moved from Espn/Common/ to Espn/Football/ so the file location
mirrors the BaseballEventCompetitionStatusDocumentProcessor at
Espn/Baseball/. Test file renamed and moved to match.

Also extracts EventCompetitionStatusProcessorBase<TDataContext> as an
intermediate base between the two sport-specific processors and
DocumentProcessorBase. Base owns the sport-agnostic lifecycle that
fires after the typed status entity has been built and the prior row
loaded:
  1. Single Competition→Contest lookup for ContestId + SeasonWeekId
  2. ContestStatusChanged publish on a transition
  3. ContestCompleted defensive publish on first observation of FINAL
  4. Contest.CancelledUtc stamp on first observation of CANCELED

Subclasses keep only the parts that genuinely differ by sport: DTO
deserialization, entity construction (FootballCompetitionStatus vs
BaseballCompetitionStatus with its MLB-only fields and FeaturedAthletes
children), typed-DbSet load + hard-replace, and the final AddAsync +
SaveChangesAsync. They call base.HandleStatusLifecycleAsync(...) once
the entity is built and the prior status's StatusTypeName is in hand.

Three lifecycle hunks that PR #428 had duplicated across the two
processors collapse to a single source. Drift risk goes to zero — the
two processors can't disagree on event payload shape, log message
text, or the CancelledUtc-as-irrevocable contract.

Sport prefixes ("MLB Contest status changed", "MLB Contest cancelled")
that the baseball variant had on its now-shared log messages are
dropped — Sport is already on the structured log fields via
command.Sport, and the shared base shouldn't carry MLB-specific
copy.

CausationId.Producer.EventCompetitionStatusDocumentProcessor (the
Guid field name in the static lookup) is intentionally NOT renamed —
that identifier is referenced by the streamer replay services and is
a stable provenance tag, not a class name. Renaming would be
incidental churn.
