+++ src/SportsData.Producer/Infrastructure/Data/Entities/ContestBase.cs
@@ -64,10 +64,11 @@ namespace SportsData.Producer.Infrastructure.Data.Entities
         public DateTime? FinalizedUtc { get; set; }
 
         // === Cancellation ===
-        // Stamped (once) by EventCompetitionStatusDocumentProcessor when
-        // ESPN reports StatusTypeName == "STATUS_CANCELED". Treated as
-        // terminal/irrevocable — ContestEnrichmentJob excludes these from
-        // future runs. See docs/contest-enrichment-historical-sweep.md.
+        // Stamped (once) by EventCompetitionStatusProcessorBase (the shared
+        // lifecycle hook used by both Football + Baseball status-doc
+        // processors) when ESPN reports StatusTypeName == "STATUS_CANCELED".
+        // Treated as terminal/irrevocable — ContestEnrichmentJob excludes
+        // these from future runs. See docs/contest-enrichment-historical-sweep.md.
         public DateTime? CancelledUtc { get; set; }
 
         // === Helpers (not mapped to DB) ===
